requirements.txt
pytest
requests
jsonschema
selene==2.0.0rc7
allure-pytest
allure-python-commons~=2.13.5